Conflict resolutions:
- tsconfig.json: drop unused paths (per main's TS6 upgrade), keep
  lib/target es2022 (per vite migration)
- package.json: take main's newer versions (electron 41.5.0,
  typescript 6.0.3, etc.); drop webpack-only deps; keep vite 6.3.5
  and plugin-vite
- package-lock.json: regenerated against merged package.json
- forge.config.js: keep vite plugin config; add whats-new-preload to
  build entries
- splash-screen.ts: keep getRendererUrl (vite); drop the
  setIgnoreMouseEvents workaround now that electron 41.5.0 ships
  the upstream fix

Wire What's New (added in main) into the vite build:
- vite.whats-new-preload.config.ts: new preload bundle config
- vite.renderer.config.ts: add whats-new HTML to renderer inputs
- whats-new-host.html: add module script tag for vite entry discovery
- whats-new-window.ts: replace WHATS_NEW_*_WEBPACK_ENTRY constants
  with getRendererUrl/getPreloadPath helpers

+++ src/cpp/session/include/session/SessionOptions.hpp
@@ -222,7 +222,7 @@ public:
       initialProjectPath_.clear();
    }
 
-   int limitRpcClientUid() const
+   int64_t limitRpcClientUid() const
    {
       return limitRpcClientUid_;
    }
@@ -279,7 +279,7 @@ private:
    core::FilePath resourcePath_;
 
    unsigned int authMinimumUserId_;
-   int limitRpcClientUid_;
+   int64_t limitRpcClientUid_;
 
    // in-session generated RSA keys
    std::string sessionRsaPublicKey_;
